Insert bitcasts if the actual type of a declared function trivially differs from its declared type.
(At least, that's the idea; for now it only applies to one specific function.)

This is a fix/workaround for the issue of passing Foster types (e.g. array
types) to C code, which has no way of naming/describing the LLVM-level
translations of the Foster types. Thus, if we see such a mismatch, we'll
insert a bitcast.

It's unclear so far what the best long-term strategy for C interop will be.