Fix a lifecycle/invariant bug which resulted in occasional double-allocation.
Turned out to (essentially) be one missing line: we weren't clearing the
local frame15 allocator when initiaing a GC. This meant that, under certain
conditions, the same frame15 would exist in the local allocator and the
clean_frame15s list. The result would be the same frame getting allocated
into twice.