Contify more exhaustively.
In particular, when we contify or inline a function, substituting for the return
continuation can produce new opportunities for contification, which we were
(until now) ignoring.

This diff also includes preliminary code for contification of recursive nests,
although it turns out the rest of the code sufficed to handle the cases seen
so far.

With this change, we also enable forced contification of blocks in C2F.