Implement default-dependent subheaps.
This provides a way of eliding the cost of maintaining the remembered set for
temporary data. The expectation for temporary data is that it is likely to be
deleted before the default subheap is collected, and if that assumption proves
false, we can just collect the temporary subheap along with the default subheap.