Begin implementing lightweight compiler and runtime support for tagged & word-packed arbitrary-precision Int values.
What's fun is that almost all the needed logic is written in Foster library code.

There are a few primitives for constructing and destructing Ints, which compile to just a few instructions each.
These primitives are wrapped by library code; all the compiler does is emit calls to the appropriate library wrappers.

One primitive (tagged Int to bignum) can potentially undermine type safety. Soon it will be protected by a refinement type which ensures safety by static evidence passing.

Currently the compiler does not forcibly/silently import any code to provide the needed symbols; the idea, for now, is that the user will do so explicitly. In theory, the user could provide an alternative implementation. Longer term, the compiler should probably import a default module automatically, with the potential for an explicit override.