Switch from ansi-wl-pprint to prettyprinter.
The biggest fallout from this change is the introduction of the PrettyT
typeclass. In prettyprinter, the Doc type is parameterized by an
annotation type (such as terminal coloring information), and the Pretty
typeclass is *parametric* over the annotation type. This implies that
*instances* of Pretty cannot use concrete annotation types (...such
as terminal coloring...). So we introduce the PrettyT typeclass for
instances which might contain terminal coloring.

It's an interesting case study for both the advantages and disadvantages
of using types to encode static knowledge. On the one hand, it's nice to
be able to tell at a call site whether it will produce formatted output
or plain output. However, using two type classes forces the programmer
to manually propagate constraints through the call graph: if a leaf node
produces formatted output, its caller must change its type signatures
(and switch from calling pretty to use prettyT instead), recursively.
Also, relying on the type system is somewhat awkward. Type error messages
for mismatches aren't particularly great: GHC describes unification
failures on rigid type variables. (The underlying problem could perhaps
be described more intuitively as "[pretty's type class constraint signature]
must be able to work for any type instantiation of [the annotation type
parameter], but [code that should be using prettyT] only works if the
type parameter is instantiated to [AnsiStyle].")

It would be interesting to compare the advantages and disadvantages of
encoding this knowledge via effects rather than type parameterization
of data types and type class signatures.