Add expected value to prim __COMPILES__
In order for dynamic semantics to not depend on static semantics,
we need a value for the dynamic semantics when the static semantics goes
un-evaluated.

Alternatively, instead of evaluating to a boolean, the primitive could
be made into a non-compilability assertion: it would evaluate to unit,
but hard-fail compilation if the snippet would typecheck without issue.